Update quickjs-libc.c - finalized change
after running executable files (and not just the test_std.js dos commands) I verified that as much functionality is the same as possible between linux and windows on all my platforms. 
CreateProcess with command line is tricky but it's needed to emulate the linux functionality.
I think this file change is final.  sorry that it wasn't final the first time.