Update quickjs-libc.c
this is the version with inherit defaults to false (silent). 
pass one pipe or pass inherit=true and the child inherits stdio by default.
I have the other version backed up, but test_os_exec.js is updated to demonstrate that shell text is not passed unless the inherit flag is set. It seems like the right thing. linux code is also updated.