Update test_std.js
updated to match, moved win32 admin flag to top with comment, removed lstat in win32. win32 has no lstat.

I can't find the right header online and it's in the microsoft hardware package because it's a device driver level change. That takes apart the node on the drive and determines its location. 
I can detect that it's a true symlink with a very small piece of code, but it doesn't tell me anything about the path. The path can be broken in so many different ways and we'd have no idea.