corrects cwd from 3 places you can execute script
test262 executes from the root.
I have to have the location of the tests folder.
I am currently using getcwd and testing cwd, which is something that needed to be done.
Executing manually requires assert, which is located in /tests, but it is possible someone executes test262 in a build folder. I don't know what to do then!